+++ src/server/plugins/typeorm.ts
@@ -1,23 +1,9 @@
 import { FastifyPluginAsync } from "fastify";
 import fp from "fastify-plugin";
-import { Repository } from "typeorm";
 
 import { AppDataSource } from "../../data/data-source";
 import { Account } from "../../data/entity/account";
 import { Person } from "../../data/entity/person";
-
-declare module "fastify" {
-  interface FastifyInstance {
-    db: {
-      accountRepository: Repository<Account>;
-      personRepository: Repository<Person>;
-    };
-  }
-  interface FastifyRequest {
-    resolvedPerson?: Person; // Optional resolved person for account creation
-    personId?: number; // Optional foreign key ID for the Person entity
-  }
-}
 
 const typeormPlugin: FastifyPluginAsync = async (fastify, opts) => {
   try {
